+++ src/host/handlers/update-configuration.ts
@@ -1,12 +1,16 @@
 import { IRequestHandler } from "@/common/messaging/core/types";
 import * as vscode from "vscode";
+import nugetApiFactory from "../nuget/api-factory";
+
 export default class UpdateConfiguration
   implements IRequestHandler<UpdateConfigurationRequest, UpdateConfigurationResponse>
 {
   async HandleAsync(request: UpdateConfigurationRequest): Promise<UpdateConfigurationResponse> {
     let config = vscode.workspace.getConfiguration("NugetGallery");
 
-    let sources = request.Configuration.Sources.map((x) => JSON.stringify({ name: x.Name, url: x.Url }));
+    let sources = request.Configuration.Sources
+      .filter((x) => !x.IsReadOnly)
+      .map((x) => JSON.stringify({ name: x.Name, url: x.Url }));
 
     await config.update(
       "credentialProviderFolder",
@@ -15,6 +19,8 @@ export default class UpdateConfiguration
     );
     await config.update("skipRestore", request.Configuration.SkipRestore, vscode.ConfigurationTarget.Global);
     await config.update("sources", sources, vscode.ConfigurationTarget.Global);
+
+    nugetApiFactory.Invalidate();
     return {};
   }
 }